+++ tests/__init__.py
@@ -1,0 +1,4 @@
+import sublime_plugin
+
+
+sublime_plugin.reload_plugin('phpunitkit.tests.helpers')